backend/requirements.txt
django
djangorestframework
django-cors-headers
pylint
pylint_django 